requirements.txt
# File: requirements.txt
requests==2.31.0
boto3==1.29.0
python-dotenv==1.0.0
pandas==2.1.3
polars==0.19.0
pyarrow==14.0.0
deltalake==0.13.0
pytz==2023.3
tqdm==4.66.0
